safari-534.53-branch/LayoutTests/fast/dom/Geolocation/reentrant-error-expected.txt

Tests that reentrant calls to Geolocation methods from the error callback are OK.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is mockCode
PASS error.message is mockMessage

PASS error.code is mockCode
PASS error.message is mockMessage
PASS successfullyParsed is true

TEST COMPLETE
